+++ utils/parallel.py
@@ -44,16 +44,16 @@ def find_facilities_for_clients(clients):
 
 
 _ELEMENTS: list | None = None
-_SIM_FN: Callable[[Any, Any], int] | None = None
+_SIM_FN: Callable[[Any, Any], float] | None = None
 
-def init_similarity(elements, sim_fn: Callable[[Any, Any], int]):
+def init_similarity(elements, sim_fn: Callable[[Any, Any], float]):
     global _ELEMENTS, _SIM_FN
     _ELEMENTS = elements
     _SIM_FN = sim_fn
 
 def compute_similarity_row(i: int):
     n = len(_ELEMENTS)
-    row = np.zeros(n, dtype=int)
+    row = np.zeros(n, dtype=float)
     element_i = _ELEMENTS[i]
     for j in range(i + 1, n):
         row[j] = _SIM_FN(element_i, _ELEMENTS[j])
